Develop in a regular spreadsheet
An addin doesn't appear in Workbooks anymore. So I needed some more
code to manually trigger the export, but then I couldn't deploy it
directly since it depended on a reference regular users don't have
(they don't need to).

It's way easier to develop a simple workbook, then save it as an addin
when I want to deploy it.